# api/src/openapi.yaml
openapi: 3.1.0
info:
  title: Deffavia API
  version: "1.1.0"
  description: API para reservas de asientos (orden + ítems), exportación/importación XML, reportes y correo.
servers:
  - url: http://localhost:3001
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        ok: { type: boolean, example: false }
        error: { type: string }
...
    Seat:
      type: object
      properties:
        id: { type: integer }
        code: { type: string }
        class: { type: string, enum: [business, economy] }
        seat_status: { type: string, description: "Estado en vista v_seat_occupancy (p.ej. 'occupied' | 'free')" }
...
    RecommendationsResponse:
      type: object
      properties:
        ok: { type: boolean }
        recommendations:
          type: object
          properties:
            favoriteClass: { type: string, enum: [business, economy], nullable: true }
            favoriteSeats:
              type: array
              items:
                type: object
                properties:
                  code: { type: string }
                  class: { type: string, enum: [business, economy] }
                  timesReserved: { type: integer }
            suggestedSeats:
              type: array
              items:
                type: object
                properties:
                  code: { type: string }
                  class: { type: string, enum: [business, economy] }
                  available: { type: boolean }
            popularSeats:
              type: array
              items:
                type: object
                properties:
                  code: { type: string }
                  class: { type: string, enum: [business, economy] }
                  totalReservations: { type: integer }
            reasoning: { type: string }
...
paths:
...
  /api/seats:
    get:
      tags: [Seats]
      summary: Asientos con estado de ocupación
      parameters:
        - in: query
          name: class
          schema: { type: string, enum: [business, economy] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Seat' }
...
  /api/recommendations/seats:
    get:
      tags: [Recommendations]
      summary: Recomendaciones de asientos basadas en historial
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: userEmail
          required: false
          description: Si no se envía, se usa el email del token
          schema: { type: string, format: email }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecommendationsResponse' }
        '400':
          description: Error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...